Menambah role dan meredirect user ber role superuser ke tempatnya

+++ app/Http/Controllers/SuperUser/UsersController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\SuperUser;
 use App\Http\Controllers\Controller;
 use App\User;
 use App\Role;
+use Gate;
 use Illuminate\Http\Request;
 
 class UsersController extends Controller
@@ -63,6 +64,9 @@ class UsersController extends Controller
      */
     public function edit(User $user)
     {
+        if(Gate::denies('edit-users')){
+          return redirect(route('superuser.users.index'));
+        }
         $roles = Role::all();
 
         return view('superuser.users.edit')->with([
@@ -82,6 +86,11 @@ class UsersController extends Controller
     {
         $user->roles()->sync($request->roles);
 
+        $user->nama = $request->nama;
+        $user->email = $request->email;
+        if($user->save())$request->session()->flash('success', $user->nama. ' has been updated');
+        else $request->session()->flash('error', 'User could not be updated');
+
         return redirect()->route('superuser.users.index');
     }
 
@@ -93,6 +102,10 @@ class UsersController extends Controller
      */
     public function destroy(User $user)
     {
+      if(Gate::denies('delete-users')){
+        return redirect(route('superuser.users.index'));
+      }
+
       $user->roles()->detach();
       $user->delete();
